+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+# v0.1.0
+
+- Add mapToObj helper
+- Following SEMVER from this version
+
 # v0.0.10
 
 - Fixed bug in hot-swaping related to edge case in mergeStates